fct_extract/requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
packaging==23.2
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.2.0
